I think it's Ok..

+++ get_next_line_utils.c
@@ -59,10 +59,10 @@ void	*ft_memset(void *s, int c, size_t n)
 char	*ft_strjoin(char *s1, char *s2)
 {
 	char	*newstr;
-	size_t	lens1;
-	size_t	lens2;
-	size_t	i;
-	size_t	j;
+	int		lens1;
+	int		lens2;
+	int		i;
+	int		j;
 
 	if (!s1 || !s2)
 		return (NULL);
